Pre release after sincup

+++ mohican.sh
@@ -14,11 +14,18 @@ get_pid() {
   PID_MASTER_PROCESS=$(head -n 1 "$PID_FILE")
 }
 
+
 get_has_server_started() {
   HAS_SERVER_STARTED=$(ps aux | grep ./mohican.out | wc -l)
 }
 
+
 start() {
+  if [ ! -d cmake-build-debug ]; then
+    echo "To start server you need to make 'sudo ./mohican.sh build'"
+    exit 1
+  fi
+
   get_has_server_started
   if [ "$HAS_SERVER_STARTED" \> 1 ]; then
     echo "Server has already started!"
@@ -33,13 +40,16 @@ start() {
     touch access.log
     touch error.log
     echo "Starting $SERVER_NAME Server..."
-    rm mohican.out
+    if [ -f mohican.out ]; then
+      rm mohican.out
+    fi
     cp ./cmake-build-debug/mohican.out mohican.out
     "$MOHICANS_HOME"/mohican.out
     echo "Server started!"
     exit 0
   fi
 }
+
 
 stop_soft() {
   get_has_server_started
@@ -55,6 +65,7 @@ stop_soft() {
 	fi
 }
 
+
 stop_hard() {
   get_has_server_started
   if [ ! "$HAS_SERVER_STARTED" \> 1 ]; then
@@ -68,6 +79,7 @@ stop_hard() {
     exit 0
   fi
 }
+
 
 reload_soft() {
   get_has_server_started
@@ -83,6 +95,7 @@ reload_soft() {
   fi
 }
 
+
 reload_hard() {
   get_has_server_started
   if [ ! "$HAS_SERVER_STARTED" \> 1 ]; then
@@ -97,6 +110,7 @@ reload_hard() {
   fi
 }
 
+
 status() {
   get_has_server_started
   if [ "$HAS_SERVER_STARTED" \> 1 ]; then
@@ -106,45 +120,93 @@ status() {
   fi
 }
 
+
 create_config() {
   cp "$DEFAULT_PATH_TO_CONFIG" settings/mohican.conf
 }
+
+
+build() {
+  if [ -d cmake-build-debug ]; then
+    cd cmake-build-debug
+    make clean && make
+    cd ..
+  else
+    mkdir cmake-build-debug
+    cd cmake-build-debug
+    cmake ..
+    make clean && make
+    cd ..
+  fi
+}
+
 
 case $1 in
   start)
     start
   ;;
+
   stop)
-  shift
-  case $1 in
-    soft)
-      stop_soft
-      exit 0
-    ;;
-    hard)
-      stop_hard
-      exit 0
-    ;;
-  esac
-  echo "Usage : <hard|soft>";
+    shift
+    case $1 in
+      soft)
+        stop_soft
+      ;;
+      hard)
+        stop_hard
+      ;;
+    esac
+    echo "Usage : <hard|soft>";
   ;;
+
   reload)
-  shift
-  case $1 in
-    soft)
-      reload_soft
-      exit 0
-    ;;
-    hard)
-      reload_hard
-      exit 0
-    ;;
-  esac
-  echo "Usage : <hard|soft>";
+    shift
+    case $1 in
+      soft)
+        reload_soft
+      ;;
+      hard)
+        reload_hard
+      ;;
+    esac
+    echo "Usage : <hard|soft>";
   ;;
+
   status)
     status
   ;;
+
+  build)
+    build
+  ;;
+
+  help)
+    shift
+    case $1 in
+      start)
+        cat .help/.start.txt
+      ;;
+      stop)
+        cat .help/.stop.txt
+      ;;
+      reload)
+        cat .help/.reload.txt
+      ;;
+      status)
+        cat .help/.status.txt
+      ;;
+      build)
+        cat .help/.build.txt
+      ;;
+      create)
+        cat .help/.create.txt
+      ;;
+      *)
+        echo "Usage : <start|stop|reload|status|build|create>";
+      ;;
+    esac
+  ;;
+
   create)
     shift
     if [ "$1" =  config ]; then
@@ -156,9 +218,9 @@ case $1 in
 
   *)
     if [ "$(ps aux | grep ./mohican.out | wc -l)" \> 1 ]; then
-      echo "Usage : <stop|reload|status>";
+      echo "Usage : <stop|reload|status|help>";
     else
-      echo "Usage : <start|status|create>";
+      echo "Usage : <start|build|status|create|help>";
     fi
   ;;
 esac
